PyModule/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.30,<1.0
aiohttp>=3.9,<4.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
pandas>=2.0,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.4,<2.0
xgboost>=2.0,<4.0
joblib>=1.3,<2.0
python-docx>=1.1,<2.0
python-dotenv>=1.0,<2.0
pywin32>=306; platform_system=="Windows"
python-escpos>=3.1,<4.0
pyusb>=1.2,<2.0
